Add auth support for UAA IDP
A fair bit of code refactoring to support multiple IDPs,
include new support for UAA.

IDP used in CLI context creation is stored in the Context's
additionalMetadata map.

`context get-token` is updated to refreshed token based on the
IDP associated with said context

tanzu context create / tanzu login
are update to use UAA IDP unless the CLI recognizes the endpoint
as a public Tanzu Platform service endpoint or if CSP idp is explicitly
requested with the --force-csp hidden flag.
