+++ pkg/auth/uaa/doc.go
@@ -1,0 +1,5 @@
+// Copyright 2022 VMware, Inc. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+// Package uaa provides UAA token utilities.
+package uaa
