+++ pkg/auth/uaa/tanzu_test.go
@@ -1,0 +1,4 @@
+// Copyright 2023 VMware, Inc. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package uaa
