+++ pkg/command/login_helper.go
@@ -93,13 +93,22 @@ func configureTanzuPlatformServiceEndpointsForSM(tpEndpoint string) error {
 	// tanzu platform url used for UI directly. In this can remove `www.` prefix.
 	u.Host = strings.TrimPrefix(u.Host, "www.")
 
-	tanzuHubEndpoint = fmt.Sprintf("%s://api.%s/hub", u.Scheme, u.Host)
-	tanzuTMCEndpoint = fmt.Sprintf("%s://ops.%s", u.Scheme, u.Host)
-	tanzuUCPEndpoint = fmt.Sprintf("%s://ucp.%s", u.Scheme, u.Host)
+	tanzuHubEndpoint = fmt.Sprintf("%s://%s/hub/hub", u.Scheme, u.Host)
+	tanzuTMCEndpoint = fmt.Sprintf("%s://%s/ops", u.Scheme, u.Host)
+	tanzuUCPEndpoint = fmt.Sprintf("%s://%s/ucp", u.Scheme, u.Host)
+	tanzuAuthEndpoint = fmt.Sprintf("%s://%s/auth", u.Scheme, u.Host)
+	if strings.Contains(u.Host, "localhost") || strings.Contains(u.Host, "127.0.0.1") {
+		tanzuAuthEndpoint = fmt.Sprintf("%s://%s", u.Scheme, u.Host)
+	}
+
 	return nil
 }
 
 func isTanzuPlatformSaaSEndpoint(tpEndpoint string) bool {
+	if forceCSP {
+		return true
+	}
+
 	if tpEndpoint == "" {
 		return false
 	}
