+++ pkg/command/org_validation_helper.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/vmware-tanzu/tanzu-plugin-runtime/log"
 
-	"github.com/vmware-tanzu/tanzu-cli/pkg/auth/csp"
+	"github.com/vmware-tanzu/tanzu-cli/pkg/auth/common"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/centralconfig"
 	"github.com/vmware-tanzu/tanzu-cli/pkg/constants"
 )
@@ -30,7 +30,7 @@ type tapScopesGetter func() ([]string, error)
 // the central configuration. If the central configuration doesn't have any Tanzu Platform for Kubernetes scopes listed, it will return success.
 // It will skip the validation and return success if TANZU_CLI_SKIP_TANZU_CONTEXT_TAP_SCOPES_VALIDATION environment
 // variable is set to true
-func validateTokenForTAPScopes(claims *csp.Claims, scopesGetter tapScopesGetter) (bool, error) {
+func validateTokenForTAPScopes(claims *common.Claims, scopesGetter tapScopesGetter) (bool, error) {
 	if skipTAPScopeValidation, _ := strconv.ParseBool(os.Getenv(constants.SkipTAPScopesValidationOnTanzuContext)); skipTAPScopeValidation {
 		return true, nil
 	}
