+++ pkg/command/org_validation_helper_test.go
@@ -7,27 +7,27 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/vmware-tanzu/tanzu-cli/pkg/auth/csp"
+	"github.com/vmware-tanzu/tanzu-cli/pkg/auth/common"
 )
 
 func TestValidateTokenForTanzuScopes(t *testing.T) {
 	tests := []struct {
 		name         string
-		claims       *csp.Claims
+		claims       *common.Claims
 		scopesGetter tapScopesGetter
 		expectedBool bool
 		expectedErr  error
 	}{
 		{
 			name:         "When the central config does not have any Tanzu Platform for Kubernetes scopes listed, it should return success.",
-			claims:       &csp.Claims{Permissions: []string{}},
+			claims:       &common.Claims{Permissions: []string{}},
 			scopesGetter: func() ([]string, error) { return []string{}, nil },
 			expectedBool: true,
 			expectedErr:  nil,
 		},
 		{
 			name:   "When the token has at least one of the specified Tanzu Platform for Kubernetes scopes listed in central config, it should return success",
-			claims: &csp.Claims{Permissions: []string{"external/UID-123-567/matching-scope", "csp:org_admin", "csp:developer"}},
+			claims: &common.Claims{Permissions: []string{"external/UID-123-567/matching-scope", "csp:org_admin", "csp:developer"}},
 			scopesGetter: func() ([]string, error) {
 				return []string{"matching-scope", "matching-another-scope"}, nil
 			},
@@ -36,7 +36,7 @@ func TestValidateTokenForTanzuScopes(t *testing.T) {
 		},
 		{
 			name:   "When the token lacks at least one of the specified Tanzu Platform for Kubernetes scopes listed in the central config, it should return an error",
-			claims: &csp.Claims{Permissions: []string{"external/UID-123-567/non-matching-scope", "csp:org_member", "csp:software_installer"}},
+			claims: &common.Claims{Permissions: []string{"external/UID-123-567/non-matching-scope", "csp:org_member", "csp:software_installer"}},
 			scopesGetter: func() ([]string, error) {
 				return []string{"matching-scope", "matching-another-scope"}, nil
 			},
@@ -45,7 +45,7 @@ func TestValidateTokenForTanzuScopes(t *testing.T) {
 		},
 		{
 			name:   "It should return an error if fetching the Tanzu Platform for Kubernetes scopes from the central config fails",
-			claims: &csp.Claims{},
+			claims: &common.Claims{},
 			scopesGetter: func() ([]string, error) {
 				return nil, errors.New("error retrieving Tanzu Platform for Kubernetes scopes")
 			},
